fix: resolve comments regarding api clients and utilities
