--- nisystemlink/utilities/data_exporter/_api_clients/_constants/__init__.py
@@ -1,1 +1,0 @@
-# flake8: noqa
