--- nisystemlink/utilities/data_exporter/_api_clients/_constants/_base_http_routes.py
@@ -1,3 +1,0 @@
-class BaseHttpRoutes:
-    WEBAPPS_BASE_ROUTE = "/niapp/v1/webapps"
-    TEST_MONITOR_BASE_ROUTE = "/nitestmonitor/v2"
