--- nisystemlink/utilities/data_exporter/_api_clients/_constants/_system_link_http_keys.py
@@ -1,4 +1,0 @@
-class SystemLinkHttpKeys:
-    FILTER = "filter"
-    RESULT_FILTER = "resultFilter"
-    UPDATED_AT = "UPDATED_AT"
