--- nisystemlink/utilities/data_exporter/_api_clients/_models/_named_value.py
@@ -1,8 +1,0 @@
-from typing import Union
-
-from pydantic import BaseModel
-
-
-class NamedValue(BaseModel):
-    name: str
-    value: Union[str, float, int]
