--- nisystemlink/utilities/data_exporter/_api_clients/_models/_status.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel, Field
-
-
-class Status(BaseModel):
-    status_type: str = Field(..., alias="statusType")
-    status_name: str = Field(..., alias="statusName")
