--- nisystemlink/utilities/data_exporter/_api_clients/_models/_step_data.py
@@ -1,8 +1,0 @@
-from typing import Dict, List
-
-from pydantic import BaseModel
-
-
-class StepData(BaseModel):
-    text: str
-    parameters: List[Dict[str, str]]
