+++ nisystemlink/utilities/data_exporter/parametric_data_exporter/__init__.py
@@ -1,3 +1,4 @@
 from ._parametric_data_exporter import ParametricDataExporter
+from ._models import QueryLinqFilter
 
 # flake8: noqa
