+++ nisystemlink/utilities/data_exporter/parametric_data_exporter/_models.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class QueryLinqFilter(BaseModel):
+    filter: str
+    result_filter: str
