+++ nisystemlink/utilities/data_exporter/parametric_data_exporter/_parametric_data_exporter.py
@@ -1,33 +1,28 @@
 import io
-from typing import Any, List, Tuple
+from typing import List, Tuple
 
 from nisystemlink.utilities.data_exporter._api_clients._api_clients import ApiClients
-from nisystemlink.utilities.data_exporter._api_clients._constants._system_link_http_keys import (
-    SystemLinkHttpKeys,
-)
-from nisystemlink.utilities.data_exporter._api_clients._models._result_response import (
+from nisystemlink.utilities.data_exporter._api_clients._models import (
     ResultResponse,
-)
-from nisystemlink.utilities.data_exporter._api_clients._models._step_response import (
     StepResponse,
 )
 from nisystemlink.utilities.data_exporter.parametric_data_exporter._dataframe_handles import (
     merge_dataframes,
     normalize_dataframe,
 )
-from pydantic import parse_obj_as
+from nisystemlink.utilities.data_exporter.parametric_data_exporter._models import (
+    QueryLinqFilter,
+)
 
 
 class ParametricDataExporter:
-    __api_clients: ApiClients
-    __query_linq_filters: Any
 
-    def __init__(self, query_linq_filters: Any):
+    def __init__(self, query_linq_filter: QueryLinqFilter):
         self.__api_clients = ApiClients()
-        self.__query_linq_filters = query_linq_filters
+        self.__query_linq_filter = query_linq_filter
 
-    async def export_csv_data(self):
-        results, steps = await self.__get_results_and_steps(self.__query_linq_filters)
+    async def export_csv_data(self) -> io.BytesIO:
+        results, steps = await self.__get_results_and_steps(self.__query_linq_filter)
 
         results_df = normalize_dataframe(
             [result.dict() for result in results], "result"
@@ -60,25 +55,22 @@ class ParametricDataExporter:
         return csv_file
 
     async def __get_results_and_steps(
-        self, query_linq_filter: dict
+        self, query_linq_filter: QueryLinqFilter
     ) -> Tuple[List[ResultResponse], List[StepResponse]]:
         """Get the results and steps based on the query linq filter.
 
         Args:
-            query_linq_filter (dict): The query linq filter of the dataspace.
+            query_linq_filter (Dict[str, str]): The query linq filter of the dataspace.
 
         Returns:
             Tuple[List[Dict], List[Dict]]: The results and steps based on the query linq filter.
         """
-        steps_filter = query_linq_filter.get(SystemLinkHttpKeys.FILTER)
-        results_filter = query_linq_filter.get(SystemLinkHttpKeys.RESULT_FILTER)
+        steps_filter = query_linq_filter.filter
+        results_filter = query_linq_filter.result_filter
 
         results = await self.__api_clients.results_client.query_results(results_filter)
-        results = parse_obj_as(List[ResultResponse], results)
-
         steps = await self.__api_clients.steps_client.query_steps(
             steps_filter, results_filter
         )
-        steps = parse_obj_as(List[StepResponse], steps)
 
         return results, steps
