--- nisystemlink/utilities/data_exporter/parametric_data_exporter/configuration/constants/__init__.py
@@ -1,3 +1,0 @@
-from ._column_groups import ColumnGroup
-
-# flake8: noqa
