--- nisystemlink/utilities/data_exporter/parametric_data_exporter/configuration/constants/_column_groups.py
@@ -1,6 +1,0 @@
-class ColumnGroup:
-    PROPERTIES = "properties"
-    STATUS_TYPE_SUMMARY = "status_type_summary"
-    INPUTS = "inputs"
-    OUTPUTS = "outputs"
-    DATA = "data"
