--- nisystemlink/utilities/data_exporter/parametric_data_exporter/configuration/constants/_product_columns.py
@@ -1,14 +1,0 @@
-product_columns = [
-    "product_id",
-    "product_part_number",
-    "product_name",
-    "product_family",
-    "product_updated_at",
-    "product_keywords",
-    "product_file_ids",
-    "product_workspace",
-]
-"""
-- product columns in the dataframe (prefix added)
-- properties won't be here. we can arrange it as per our need in the reorder function
-"""
