--- nisystemlink/utilities/data_exporter/parametric_data_exporter/configuration/constants/_result_columns.py
@@ -1,21 +1,0 @@
-result_columns = [
-    "result_started_at",
-    "result_updated_at",
-    "result_program_name",
-    "result_id",
-    "result_system_id",
-    "result_host_name",
-    "result_operator",
-    "result_part_number",
-    "result_serial_number",
-    "result_total_time_in_seconds",
-    "result_keywords",
-    "result_file_ids",
-    "result_workspace",
-    "result_status.status_type",
-    "result_status.status_name",
-]
-"""
-- result columns in the dataframe (prefix added)
-- status_type_summary and properties won't be here. we can arrange them as per our need in the reorder function
-"""
