--- nisystemlink/utilities/data_exporter/parametric_data_exporter/configuration/constants/_step_columns.py
@@ -1,21 +1,0 @@
-step_columns = [
-    "step_name",
-    "step_step_type",
-    "step_step_id",
-    "step_parent_id",
-    "step_result_id",
-    "step_path",
-    "step_path_ids",
-    "step_status.status_type",
-    "step_status.status_name",
-    "step_total_time_in_seconds",
-    "step_started_at",
-    "step_updated_at",
-    "step_has_children",
-    "step_workspace",
-    "step_keywords",
-]
-"""
-- step columns in the dataframe (prefix added)
-- properties won't be here. we can arrange it as per our need in the reorder function
-"""
