+++ pyproject.toml
@@ -38,6 +38,9 @@ requests = "^2.28.1"
 uplink   = "^0.9.7"
 pydantic = "^1.10.2"
 pyyaml = "^6.0.1"
+pandas = "^2.2.3"
+aiohttp = "^3.11.12"
+backoff = "^2.2.1"
 
 [tool.poetry.group.dev.dependencies]
 black               = ">=22.10,<25.0"
